// Asset/Server/src/common/const/relation.name.enum.ts
/**
 * User class name Enum
 */
export enum RelationName {
  HAS = 'HAS',
  HAS_VIRTUAL_RELATION = 'HAS_VIRTUAL_RELATION',
  HAS_FORM = 'HAS_FORM',
  HAS_FORM_DATA = 'HAS_FORM_DATA',
  PARENT_OF = 'PARENT_OF',
  MERGEDJS = 'MERGEDJS',
  MERGEDZN = 'MERGEDZN',
  CLASSIFICATION_OF = 'CLASSIFICATION_OF',
  CLASSIFIED_BY = 'CLASSIFIED_BY',
  CREATED_BY = 'CREATED_BY',
  MANUFACTURED_BY = 'MANUFACTURED_BY',
  LOCATED_IN = 'LOCATED_IN',
  WARRANTY_GUARANTOR_PARTS = 'WARRANTY_GUARANTOR_PARTS_BY',
  WARRANTY_GUARANTOR_LABOR = 'WARRANTY_GUARANTOR_LABOR_BY',
  CREATOR_OF = 'CREATOR_OF',
  CONTAINS_COMPONENT = 'CONTAINS_COMPONENT',
  ASSET_TYPE_BY = 'ASSET_TYPE_BY',
  DURATION_UNIT_BY = 'DURATION_UNIT_BY',
  WARRANTY_DURATION_UNIT_BY = 'WARRANTY_DURATION_UNIT_BY',
  SYSTEM_OF = 'SYSTEM_OF'
}